+++ posthog/src/main/java/com/posthog/internal/PostHogQueue.kt
@@ -81,16 +81,23 @@ internal class PostHogQueue(
                     deque.add(file)
                 }
 
+                var hasPendingEvents = true
                 try {
                     val os = config.encryption?.encrypt(file.outputStream()) ?: file.outputStream()
                     os.use { theOutputStream ->
                         config.serializer.serialize(event, theOutputStream.writer().buffered())
                     }
                     config.logger.log("Queued Event ${event.event}: ${file.name}.")
-
-                    flushIfOverThreshold()
                 } catch (e: Throwable) {
+                    hasPendingEvents = false
                     config.logger.log("Event ${event.event}: ${file.name} failed to parse: $e.")
+
+                    // if for some reason the file failed to serialize, lets delete it
+                    file.deleteSafely(config)
+                }
+
+                if (hasPendingEvents) {
+                    flushIfOverThreshold()
                 }
             }
         }
@@ -163,6 +170,17 @@ internal class PostHogQueue(
         }
     }
 
+    private fun deleteFileSafely(
+        file: File,
+        throwable: Throwable? = null,
+    ) {
+        synchronized(dequeLock) {
+            deque.remove(file)
+        }
+        file.deleteSafely(config)
+        config.logger.log("File: ${file.name} failed to parse: $throwable.")
+    }
+
     @Throws(PostHogApiError::class, IOException::class)
     private fun batchEvents() {
         val files = takeFiles()
@@ -175,14 +193,12 @@ internal class PostHogQueue(
                     val event = config.serializer.deserialize<PostHogEvent?>(it.reader().buffered())
                     event?.let { theEvent ->
                         events.add(theEvent)
+                    } ?: run {
+                        deleteFileSafely(file)
                     }
                 }
             } catch (e: Throwable) {
-                synchronized(dequeLock) {
-                    deque.remove(file)
-                }
-                file.deleteSafely(config)
-                config.logger.log("File: ${file.name} failed to parse: $e.")
+                deleteFileSafely(file, e)
             }
         }